Developer Mode alpha / demo v0.1 proper
First demo version of sorts. Once enabled, refresh the page and the extension will run 9 checks on the data collected. Browsing subsequent pages will loop over all log entries and run these checks, and all subsequent pages run a 10th check to compare the value of Client ID (from ga() tracker) against that of the first page. More checks will follow.

No UI yet, the icon will be red/green/grey depending on app state (red for errors, grey disabled, green for OK). All the errors are reported in the Background Page for the extension - so this needs to be run in developer mode for the moment with the background page's console open. UI to follow next.

I've also simplified and corrected issues with Status preservation (Page Action click listener apparently doesn't fire if the extension uses a popup!), and also re-ordered a ton of the internal workings and cleaned up a number of redundant objects. I've tried to keep the data flow simple but without too many monolithic functions.